+++ projects/query-search/src/lib/directives/index.ts
@@ -1,2 +1,3 @@
 export * from './query-field.directive';
 export * from './query-fields.directive';
+export * from './options-scroll.directive';